Remove unused files
The linter setup for this repository fails for these binary files,
due to incorrectly checking them for line endings as a text file.
Since the .Net development certificate is handled in a better manner
with recent .Net releases, there is no need to have any keychain files.
If this is to be deployed on a server, then new ssl certificates will
be used anyhow.
The files are not used anywhere.