Add docker compose setup of postgresql
And in passing avoid using port 80, but rather 6080 because ports
below 1024 are restricted on mac/linux.

Use a custom port 6543 for postgres, to avoid conflicts with any
databases running locally on the host itself.

Comment out ms sql server, until it is decided to continue using it,
or not, or maybe there should be two separate docker compose
alternatives, depending on the server that should be used.

Add connection strings to use postgres