Convert all pcss files to scss
The pcss suffix was invented by this project, and hard coded
in webpack to apply postcss transformations, support the sass syntax,
i.e. css with indentation for block formatting instead of braces.

When changing the files to scss the syntax is also changed to use
braces for blocks, conforming the the scss standard.