Configure appsettings loading for dotnet ef
The `dotnet ef` command line tool must run in the Data project,
and must have a defined startup that adds settings, such as
connection string.
Adjust the path to work for `dotnet ef` when run form the
`src/nscreg.Data` directory.
This explains why there is a Startup configuration for the Data
project, even if it should never run standalone.