[dev] Bugfix path for loading
It seems npm uses the path where you run the command from.
Change to use a path relative to the current file.
In passing change to use Node 16 fs/promise for async reading
of files. This is required for Node 16, that is the currrently supported
version.
This worked if one accidentally runs with node 18.