[ui] Replace statbus logo
The new logo was AI generated from Bard with the following prompt:

"""
The name is "Statbus" and it stands for STATistical BUSiness register"
and is used by statistical offices in the world for maintaining a
business registry and produce nation level statistics about those
businesses, as well as perform business surveys.

It should work both as a small icon, and as a larger logo.

By having a common categorisation provided by Statbus, through use of
international standards, it is easier for countries to compare
themselves with others.
The is part of the government branch in each country.
Statbus is provided by the Statistical Central Bureau of Norway (SSB)
as a foreign aid to developing countries who may not afford to
otherwise.

Our motto is:
"Simple to use, simply useful."

It allows a country to start small, and then improve quality of their
work through built in guides on data quality.
Furthermore it has simple built in statistics from day one,
as well as exports of data for custom charts, and database access for
advanced use cases.
"""