[dev] Improve locale support in postgres
Prepare the ground for creating the database with various locales.
Currently the locale is hardcoded, but the required locales
are loaded in the postgres image, and there is uncommented
example code to change the language.

Also cleanup the name of the file with the database password,
to be database agnostic.

Also improve the speed and logging of the docker postgres database,
for development support only.