[ef] Fix wrong database listing for legal unit search
The StatUnitSearchView was using StatId as a primary key,
but all of the Legal Unit, Local Unit and Enterprise use the SAME
StatId, and that causes Entity Framework Core to load data incorrectly.
EF Core use a dictionary with the StatId as key, so when the same
StatId is seen again, it doesn't read the actual row, but rather
re-uses the row from the dictionary.
This leads to multiple rows, with the exact same content.