[bg] Use a temporary location for temporary files
The prior code wrote files to the same directory where the executable
is running, and that doesn't make sense running inside a docker
container.
Either we must use a proper temporary storage,
or use a persistent directory.
We use proper temporary storage for now, as that is simpler to code,
and change if applicable.