[db] Cleanup roles and names
After a meeting with Christian Bendtsen, where he misunderstood
'administrator' to be a database manager role, so I changed it to
'super_user' to make clear that it is a role related to management
of data in the system.
Furthermore, I rename 'employee' to 'restricted_user', because that
is the only use case, a user that only have access to certain data,
depending on roles granted, such as specific regions or activity
categories.