[ui] Add the Carbon UI framework
Provided by IBM, with up to date examples and code for Svelte,
and with automated generation of the Svelte compatible components.
It turns out that all others I have tried, were lacking.
Semantic UI is no longer developed.
Fomantic UI is promising, but he Svantic port is not automated
and stale and not complete, and not recent.
Bootstrap is great, but SelteBoot is not Svelte 4 compatible,
and it is not automated, so there will be a while before it
is upgraded.
Material UI as a non official Svelte port, I didn't try it yet.

Carbon had a nice tutorial with understandable theming, and
all the examples provided working code.