[db] Remove unused registration_reason
There are no sensible example registration reasons provided
anywhere in the project.
Erik Søberg and myself could not think of a reasonable list
of registration reasons.
The purpose of the whole system is to have a business register,
so that is the ultimate reason, and does not need specifying.