[db] Simplify enterprise
An enterprise is a group of either legal_units or establishments,
when an establishment does not have a legal unit.
To know how to propagate information there must be one, and only one,
primary legal_unit or primary establishment.
This means that there is no need for separate information on the
enterprise since it is derived from the primary.