Revert "[import] Add birth_date, death_date and secondary_activity_category_code"
This reverts commit 3eae712bb989f996792016cf56b984dd65d7481f.
The upload to web does not directly support empty strings for dates,
they are not treated as nulls.
Either the selection must be changed to address this, or there must be some
setting/flag in PostGREST, or there must be a temporary table with TEXT,
and later conversion to the correct type, that may fail, and become NULL
if it fails.