[ui] Fix details slots for sub paths
The slots @nav, @header and @topology did not work for details pages under /legal-units/:id/ like /legal-units/:id/contact when doing a full page load.

This was due to the fact that the @contact/page.tsx did not match the route. By renaming the @contact/default.tsx it will match all subroutes.