[db] Fix bug in enterprise summary
Due to difference in the search column.

In passing split the materialized view into a view that defines it,
and another materialized view from the defining view.
This allows using the defining view to find problems that would
lead to updates on the materialized view to fail.
In passing simplify the expression for activity_category_paths.