[db] Rename scoped_tag_... to by_tag_id_...
After discussion and testing with AI, change to
  by_tag_id
  by_tag_id_unique_ident
The `by_tag` prefix signifies that this is the same concept, and that
it is used for categorisation related to a tag.
The `by_tag_id` identifies the tag use for categorisation.
The `unique_ident` signfies that there is something unique in an
external system, since `_ident` is used for identifiers in external
systems, wile `_id` is for internal systems.