feat!: Consolidate docker compose configuration
Such that there is a single docker compose command that will manage
statbus with our own app and with supabase provided services.

Also move all configuration to the top level `.env` file.

The change of structure requires preparations for deployed instances
by consolidation of `.env` and `supabase_docker/.env`,
taking into account the rename of
as well as the rename of `PUBLIC_PORT` to `APP_BIND_ADDRESS`
and the removal of `SUPABASE_NETWORK` since the networks can be
dynamically assigned, as all used ports are explicitly set.

Use the include functionality of docker compose to build up
the singular entry point.

Also generate code to produce the required overrides of the Supabase
configuration, as well as code to update supabase.