[ui] Provide activity-categories and regions with search provider
This is a workaround due to missing region- and category names in the search result. A better solution is to include these names in the search results and render them directly from the statistical unit data.

TODOs has been added to the code to remove this workaround once the extra data is returned from the search result.