db: Rewrite statistical_unit
Instead of trying to do a clever join - use an altogether different
strategy.
Establish a list of timepoints for every unit.
A timepoint is when something changes.
Then from that list of timepoints, make a list of connected
timesegments for every unit.

Using this list of segments, make a clean join for each segment.