db: Create period_active for UI time selection
The view period_active combines the relative_period with
tags that have time information into one unified structure,
and exposes them so that the UI can use it for selection
of when to query with meaningful labels.
It remains to use the name_when_input later, when we support
input on the web, and one must know what time the input
will be valid for.