db: Remove seen_in_import_at
This method of detection with views does not work, in principle,
because updates through views will not issue an update and run
the INSTEAD of function if there is no change, and therefore, on
the second identical update, all is deleted.
This must be redone in the new batch import system that is planned.