test: Fix small history load
The small history will need to use a new batch import
with custom column names, after it is written, but make the test
work for now with helper tables for mapping, to test that the
data can load.
Also, this provides an opportunity to write the code that the
future flexible batch import needs to generate.