ui: Switch to pnpm for package management
Because it offers the `pnpm up --latest` command that upgrades
all packages, unlike npm that doesn't support it.
I also tried with `npx npm-check-updates -u`
and that does upgrade the packages, but then there are dependency
inconsistencies that makes `npm install` fail.
Instead I change to pnpm that supports upgrading directly in the too.
and now with the new `pnpm run update` command,
and that support per package dependencies without conflict.
pnpm is written in rust, and is fast, and works the same
on Mac, Windows and Linux.