db: Fix history death/birth/count and track name/addresss changes
There was a logical bug in the history counting, remedied by analysing
the timeline within each range, and picking the *last* relevant
entry and using that as a starting point for analysis.
Furthermore any change must be analysed with respect to the previous
range, so the historical periods are extended with prev_stop
as the last day of the preceeding range.

Also track changes to name/address across ranges.
Tests adjusted as required.
The example data was changed to add the death date the same day
it is valid from, as it was not known in advance about the death.